@@ -1,6 +1,6 @@
 module github.com/thuongtruong109/gouse
 
-go 1.23.0
+go 1.24
 
 require (
 	github.com/AlecAivazis/survey/v2 v2.3.7
