@@ -75,28 +75,24 @@ func GetTagName(structInstance any) []string {
 // 	return newStructValue.Interface()
 // }
 
-func MergeStruct(structs ...interface{}) interface{} {
+func MergeStruct(structs ...any) any {
 	if len(structs) < 2 {
 		panic("At least two structs are required to merge")
 	}
 
-	// Create a map to store field values
 	fieldValues := make(map[string]reflect.Value)
 
-	// Iterate over each struct
 	for _, s := range structs {
 		v := reflect.ValueOf(s)
 		t := v.Type()
 
-		// Iterate over fields of each struct
-		for i := 0; i < v.NumField(); i++ {
+		for i := range v.NumField() {
 			fieldName := t.Field(i).Name
 			fieldValues[fieldName] = v.Field(i)
 		}
 	}
 
-	// Create a new map with merged fields
-	resultMap := make(map[string]interface{})
+	resultMap := make(map[string]any)
 	for fieldName, fieldValue := range fieldValues {
 		resultMap[fieldName] = fieldValue.Interface()
 	}
@@ -104,19 +100,17 @@ func MergeStruct(structs ...interface{}) interface{} {
 	return resultMap
 }
 
-func RemoveStruct(structInstance interface{}, fields ...string) interface{} {
+func RemoveStruct(structInstance any, fields ...string) any {
 	structValue := reflect.ValueOf(structInstance)
 	structType := structValue.Type()
 
-	// Create a map to store excluded field names
 	excludedFields := make(map[string]bool)
 	for _, field := range fields {
 		excludedFields[field] = true
 	}
 
-	// Create a new list of struct fields excluding the specified fields
 	var newFields []reflect.StructField
-	for i := 0; i < structValue.NumField(); i++ {
+	for i := range structValue.NumField() {
 		fieldName := structType.Field(i).Name
 		if !excludedFields[fieldName] {
 			newFields = append(newFields, structType.Field(i))
@@ -126,7 +120,6 @@ func RemoveStruct(structInstance interface{}, fields ...string) interface{} {
 	newStructType := reflect.StructOf(newFields)
 	newStructValue := reflect.New(newStructType).Elem()
 
-	// Copy values from the original struct to the new struct
 	for _, field := range newFields {
 		fieldName := field.Name
 		value := structValue.FieldByName(fieldName).Interface()
@@ -136,22 +129,19 @@ func RemoveStruct(structInstance interface{}, fields ...string) interface{} {
 	return newStructValue.Interface()
 }
 
-func AddStruct(structInstance interface{}, fields map[string]interface{}) interface{} {
+func AddStruct(structInstance any, fields map[string]any) any {
 	structType := reflect.TypeOf(structInstance)
 
-	// Create a map to store existing field values
-	existingValues := make(map[string]interface{})
-	for i := 0; i < structType.NumField(); i++ {
+	existingValues := make(map[string]any)
+	for i := range structType.NumField() {
 		fieldName := structType.Field(i).Name
 		existingValues[fieldName] = reflect.ValueOf(structInstance).Field(i).Interface()
 	}
 
-	// Add new fields to the map
 	for fieldName, value := range fields {
 		existingValues[fieldName] = value
 	}
 
-	// Create a new struct type with the combined fields
 	newStructType := reflect.StructOf(
 		func() []reflect.StructField {
 			var fields []reflect.StructField
@@ -165,11 +155,9 @@ func AddStruct(structInstance interface{}, fields map[string]interface{}) interf
 		}(),
 	)
 
-	// Create a new struct instance of the new type
 	newStructValue := reflect.New(newStructType).Elem()
 
-	// Set field values for the new instance
-	for i := 0; i < newStructType.NumField(); i++ {
+	for i := range newStructType.NumField() {
 		fieldName := newStructType.Field(i).Name
 		fieldValue := existingValues[fieldName]
 		newStructValue.Field(i).Set(reflect.ValueOf(fieldValue))
@@ -178,11 +166,9 @@ func AddStruct(structInstance interface{}, fields map[string]interface{}) interf
 	return newStructValue.Interface()
 }
 
-func SetStruct(structInstance interface{}, fieldName string, value interface{}) {
-	// Get the reflect.Value of the struct instance
+func SetStruct(structInstance any, fieldName string, value any) {
 	structValue := reflect.ValueOf(structInstance)
 
-	// Check if the structInstance is a pointer
 	if structValue.Kind() != reflect.Ptr || structValue.IsNil() {
 		Println("Struct instance must be a non-nil pointer.")
 		return
@@ -191,16 +177,13 @@ func SetStruct(structInstance interface{}, fieldName string, value interface{}) 
 	// Dereference the pointer to get the actual struct value
 	structValue = structValue.Elem()
 
-	// Get the field by name
 	field := structValue.FieldByName(fieldName)
 
-	// Check if the field is valid and exported
 	if !field.IsValid() || !field.CanSet() {
 		Printf("Field %s is unexported or not found.\n", fieldName)
 		return
 	}
 
-	// Check if the types match before setting the value
 	if reflect.ValueOf(value).Type().AssignableTo(field.Type()) {
 		field.Set(reflect.ValueOf(value))
 	} else {
@@ -208,23 +191,20 @@ func SetStruct(structInstance interface{}, fieldName string, value interface{}) 
 	}
 }
 
-func GetStruct(structInstance interface{}, fieldName string) interface{} {
+func GetStruct(structInstance any, fieldName string) any {
 	structValue := reflect.ValueOf(structInstance)
 
-	// Get the field value
 	fieldValue := structValue.FieldByName(fieldName)
 
 	return fieldValue.Interface()
 }
 
-func CloneStruct(structInstance interface{}) interface{} {
+func CloneStruct(structInstance any) any {
 	structValue := reflect.ValueOf(structInstance)
 
-	// Create a new struct instance
 	newStructValue := reflect.New(structValue.Type()).Elem()
 
-	// Copy values from the original struct to the new struct
-	for i := 0; i < structValue.NumField(); i++ {
+	for i := range structValue.NumField() {
 		fieldValue := structValue.Field(i)
 		newStructValue.Field(i).Set(fieldValue)
 	}
@@ -232,13 +212,13 @@ func CloneStruct(structInstance interface{}) interface{} {
 	return newStructValue.Interface()
 }
 
-func HasInStruct(structInstance interface{}, fieldName string) bool {
+func HasInStruct(structInstance any, fieldName string) bool {
 	structType := reflect.TypeOf(structInstance)
 	_, ok := structType.FieldByName(fieldName)
 	return ok
 }
 
-func HasEmptyInStruct(structInstance interface{}, fieldName string) bool {
+func HasEmptyInStruct(structInstance any, fieldName string) bool {
 	structValue := reflect.ValueOf(structInstance)
 	fieldValue := structValue.FieldByName(fieldName)
 
